requirements-faust.txt
faust-streaming==0.11.3
confluent-kafka==2.3.0
python-dotenv==1.0.0